update some change

+++ database/migrations/2024_08_07_063342_create_jalans_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('kode', 15)->unique();
             $table->string('nama', 50);
-            $table->integer('panjang');
+            $table->float('panjang');
             $table->integer('lebar');
             $table->timestamps();
         });
